requirements.txt
setuptools>=65.5.1
numpy==1.24.3
Pillow==9.5.0
opencv-python-headless==4.8.1.78
streamlit==1.29.0